build(pom): 更新 Maven 配置以发布到中央仓库
- 修改 groupId 为 io.github.cookiegege
- 更新版本号为1.0
- 添加项目描述和许可证信息
- 配置 Sonatype OSSRH 发布插件
- 添加 GPG 签名插件
- 配置源代码和 Javadoc 插件
