--- src/main/java/com/cookiegege/excel/ExcelListener.java
@@ -1,14 +1,0 @@
-package com.cookiegege.excel;
-
-import com.alibaba.excel.read.listener.ReadListener;
-
-/**
- * Excel 导入监听
- *
- * @author Lion Li
- */
-public interface ExcelListener<T> extends ReadListener<T> {
-
-    ExcelResult<T> getExcelResult();
-
-}
