--- src/main/java/com/cookiegege/util/CommonValidatorUtils.java
@@ -1,29 +1,0 @@
-package com.cookiegege.util;
-
-import cn.hutool.extra.spring.SpringUtil;
-import jakarta.validation.ConstraintViolation;
-import jakarta.validation.ConstraintViolationException;
-import jakarta.validation.Validator;
-import lombok.AccessLevel;
-import lombok.NoArgsConstructor;
-
-import java.util.Set;
-
-/**
- * Validator 校验框架工具
- *
- * @author Lion Li
- */
-@NoArgsConstructor(access = AccessLevel.PRIVATE)
-public class CommonValidatorUtils {
-
-    private static final Validator VALID = SpringUtil.getBean(Validator.class);
-
-    public static <T> void validate(T object, Class<?>... groups) {
-        Set<ConstraintViolation<T>> validate = VALID.validate(object, groups);
-        if (!validate.isEmpty()) {
-            throw new ConstraintViolationException("参数校验异常", validate);
-        }
-    }
-
-}
