+++ src/main/java/io/github/cookiegege/excel/ExcelListener.java
@@ -1,0 +1,14 @@
+package io.github.cookiegege.excel;
+
+import com.alibaba.excel.read.listener.ReadListener;
+
+/**
+ * Excel 导入监听
+ *
+ * @author Lion Li
+ */
+public interface ExcelListener<T> extends ReadListener<T> {
+
+    ExcelResult<T> getExcelResult();
+
+}
