+++ src/main/java/io/github/cookiegege/strategy/CellMergeStrategy.java
@@ -1,0 +1,193 @@
+package io.github.cookiegege.strategy;
+
+import com.alibaba.excel.metadata.Head;
+import com.alibaba.excel.write.merge.AbstractMergeStrategy;
+import io.github.cookiegege.annotation.CellMerge;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.SneakyThrows;
+import lombok.extern.slf4j.Slf4j;
+import org.apache.commons.collections4.CollectionUtils;
+import org.apache.poi.ss.usermodel.Cell;
+import org.apache.poi.ss.usermodel.Sheet;
+import org.apache.poi.ss.util.CellRangeAddress;
+
+import java.lang.reflect.Field;
+import java.lang.reflect.Method;
+import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+import java.util.stream.Collectors;
+
+/**
+ * 列值重复合并策略
+ *
+ * @author Lion Li
+ */
+@AllArgsConstructor
+@Slf4j
+public class CellMergeStrategy extends AbstractMergeStrategy {
+
+    private List<?> list;
+    private boolean hasTitle;
+
+    @Override
+    protected void merge(Sheet sheet, Cell cell, Head head, Integer relativeRowIndex) {
+        int headSize = head.getHeadNameList().size();
+        List<CellRangeAddress> cellList = handleRaw(list, hasTitle, headSize);
+        List<CellRangeAddress> colCellList = handleCol(list, hasTitle, headSize);
+        cellList.addAll(colCellList);
+        // judge the list is not null
+        if (CollectionUtils.isNotEmpty(cellList)) {
+            // the judge is necessary
+            if (cell.getRowIndex() == headSize && cell.getColumnIndex() == 0) {
+                for (CellRangeAddress item : cellList) {
+                    sheet.addMergedRegion(item);
+                }
+            }
+        }
+    }
+
+    @SneakyThrows
+    private static List<CellRangeAddress> handleCol(List<?> list, boolean hasTitle, Integer headSize) {
+        List<CellRangeAddress> cellList = new ArrayList<>();
+        if (CollectionUtils.isEmpty(list)) {
+            return cellList;
+        }
+        Class<?> clazz = list.get(0).getClass();
+        Field[] fields = clazz.getDeclaredFields();
+        // 有注解的字段
+        HashMap<Integer, Field> fieldHashMap = new HashMap<>();
+        for (int i = 0; i < fields.length; i++) {
+            Field field = fields[i];
+            if (field.isAnnotationPresent(CellMerge.class)) {
+                CellMerge cm = field.getAnnotation(CellMerge.class);
+                fieldHashMap.put(cm.index(), field);
+            }
+        }
+        List<Integer> sortIndexList = fieldHashMap.keySet().stream().sorted().collect(Collectors.toList());
+
+        int rowIndex = hasTitle ? headSize : 0;
+        Map<Object, RepeatCell> map = new HashMap<>();
+        // 生成两两合并单元格
+        for (int i = 0; i < list.size(); i++) {
+            for (int j = 0; j < sortIndexList.size(); j++) {
+                Integer index = sortIndexList.get(j);
+                Field field = fieldHashMap.get(index);
+                CellMerge cm = field.getAnnotation(CellMerge.class);
+
+                String name = field.getName();
+                String methodName = "get" + name.substring(0, 1).toUpperCase() + name.substring(1);
+                Method readMethod = clazz.getMethod(methodName);
+                Object rowData = list.get(i);
+                Object val = readMethod.invoke(rowData);
+                int colNum = index;
+
+                if (!cm.mergeCol()) {
+                    // 需要判断是否需要合并到这里
+                    if (map.containsKey(rowData)) {
+                        RepeatCell repeatCell = map.get(rowData);
+                        Object cellValue = repeatCell.getValue();
+                        if (!cellValue.equals(val)) {
+                            if (j - repeatCell.getCurrent() > 1) {
+                                cellList.add(new CellRangeAddress(i + rowIndex, i + rowIndex, repeatCell.getCurrent(), j));
+                            }
+                            map.put(rowData, new RepeatCell(val, colNum));
+                        } else if (j == sortIndexList.size() - 1) {
+                            if (j > repeatCell.getCurrent()) {
+                                cellList.add(new CellRangeAddress(i + rowIndex, i + rowIndex, repeatCell.getCurrent(), j));
+                            }
+                        }
+                    }
+                    continue;
+                }
+
+                if (!map.containsKey(rowData)) {
+                    map.put(rowData, new RepeatCell(val, colNum));
+                } else {
+                    RepeatCell repeatCell = map.get(rowData);
+                    Object cellValue = repeatCell.getValue();
+                    if (!cellValue.equals(val)) {
+                        if (j - repeatCell.getCurrent() > 1) {
+                            cellList.add(new CellRangeAddress(i + rowIndex, i + rowIndex, repeatCell.getCurrent(), j));
+                        }
+                        map.put(rowData, new RepeatCell(val, colNum));
+                    } else if (j == sortIndexList.size() - 1) {
+                        if (j > repeatCell.getCurrent()) {
+                            cellList.add(new CellRangeAddress(i + rowIndex, i + rowIndex, repeatCell.getCurrent(), j));
+                        }
+                    }
+                }
+            }
+        }
+        return cellList;
+    }
+
+    @SneakyThrows
+    private static List<CellRangeAddress> handleRaw(List<?> list, boolean hasTitle, Integer headSize) {
+        List<CellRangeAddress> cellList = new ArrayList<>();
+        if (CollectionUtils.isEmpty(list)) {
+            return cellList;
+        }
+        Class<?> clazz = list.get(0).getClass();
+        Field[] fields = clazz.getDeclaredFields();
+        // 有注解的字段
+        List<Field> mergeFields = new ArrayList<>();
+        List<Integer> mergeFieldsIndex = new ArrayList<>();
+        for (int i = 0; i < fields.length; i++) {
+            Field field = fields[i];
+            if (field.isAnnotationPresent(CellMerge.class)) {
+                CellMerge cm = field.getAnnotation(CellMerge.class);
+                mergeFields.add(field);
+                mergeFieldsIndex.add(cm.index() == -1 ? i : cm.index());
+            }
+        }
+        // 行合并开始下标
+        int rowIndex = hasTitle ? headSize : 0;
+        Map<Field, RepeatCell> map = new HashMap<>();
+        // 生成两两合并单元格
+        for (int i = 0; i < list.size(); i++) {
+            for (int j = 0; j < mergeFields.size(); j++) {
+                Field field = mergeFields.get(j);
+                String name = field.getName();
+                String methodName = "get" + name.substring(0, 1).toUpperCase() + name.substring(1);
+                Method readMethod = clazz.getMethod(methodName);
+                Object val = readMethod.invoke(list.get(i));
+
+                int colNum = mergeFieldsIndex.get(j);
+                if (!map.containsKey(field)) {
+                    map.put(field, new RepeatCell(val, i));
+                } else {
+                    RepeatCell repeatCell = map.get(field);
+                    Object cellValue = repeatCell.getValue();
+                    if (cellValue == null || "".equals(cellValue)) {
+                        // 空值跳过不合并
+                        continue;
+                    }
+                    if (!cellValue.equals(val)) {
+                        if (i - repeatCell.getCurrent() > 1) {
+                            cellList.add(new CellRangeAddress(repeatCell.getCurrent() + rowIndex, i + rowIndex - 1, colNum, colNum));
+                        }
+                        map.put(field, new RepeatCell(val, i));
+                    } else if (i == list.size() - 1) {
+                        if (i > repeatCell.getCurrent()) {
+                            cellList.add(new CellRangeAddress(repeatCell.getCurrent() + rowIndex, i + rowIndex, colNum, colNum));
+                        }
+                    }
+                }
+            }
+        }
+        return cellList;
+    }
+
+    @Data
+    @AllArgsConstructor
+    static class RepeatCell {
+
+        private Object value;
+
+        private int current;
+
+    }
+}
