--- src/test/java/com/cookiegege/entity/DemoData.java
@@ -1,30 +1,0 @@
-package com.cookiegege.entity;
-
-import com.alibaba.excel.annotation.ExcelIgnore;
-import com.alibaba.excel.annotation.ExcelProperty;
-import lombok.EqualsAndHashCode;
-import lombok.Getter;
-import lombok.Setter;
-
-import java.util.Date;
-
-/**
- * @author JoSuper
- * @date 2025/9/5 10:47
- */
-@Getter
-@Setter
-@EqualsAndHashCode
-public class DemoData {
-    @ExcelProperty("字符串标题")
-    private String string;
-    @ExcelProperty("日期标题")
-    private Date date;
-    @ExcelProperty("数字标题")
-    private Double doubleData;
-    /**
-     * 忽略这个字段
-     */
-    @ExcelIgnore
-    private String ignore;
-}
